requirements.txt
asgiref==3.4.1
Django==3.2.5
django-admin-rangefilter==0.8.1
pytz==2021.1
sqlparse==0.4.2
openpyxl==3.0.9
djangorestframework==3.13.1
Markdown==3.3.7
django-filter==21.1